"Add defensive check in user delete, and prevent multiple registration form submissions"
In the DeleteUserCommandHandler, added null-check for user object and associated player before removing associated EloHistories, preventing possible NullReferenceException if user or player not found.

In ExternalLogin.cshtml, added a script to hide the submit button after click, preventing accidentally multiple form submissions during registration.

+++ Server/Handlers/Commands/User/DeleteUser/DeleteUserCommandHandler.cs
@@ -19,9 +19,13 @@ public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommandRequest
             .Include(x => x.Player)
             .ThenInclude(x => x.EloHistories)
             .SingleAsync(x => x.Id == request.UserId, cancellationToken: cancellationToken);
+
+        if (user?.Player != null)
+        {
+            _dbContext.EloHistories.RemoveRange(user.Player.EloHistories);
+            await _dbContext.SaveChangesAsync(cancellationToken);
+        }
         
-        _dbContext.EloHistories.RemoveRange(user.Player.EloHistories);
-        await _dbContext.SaveChangesAsync(cancellationToken);
         await _dbContext.Users
             .Where(x => x.Id == request.UserId)
             .ExecuteDeleteAsync(cancellationToken: cancellationToken);
